# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Solana multi-wallet panel API
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
  rpc: string;
  jupiterConfigured: boolean;
}

export interface ErrorResponse {
  error: string;
  /** @nullable */
  code?: number | null;
  /** @nullable */
  errorCode?: string | number | null;
  /** @nullable */
  router?: string | null;
  /** @nullable */
  signature?: string | null;
  /** @nullable */
  status?: string | null;
}

export interface Balance {
  address: string;
  lamports: number;
  sol: number;
}

export interface MintInfo {
  mint: string;
  valid: boolean;
  decimals: number;
  supply: string;
  program: string;
}

export interface OrderInput {
  wallet: string;
  outputMint: string;
  /** @exclusiveMinimum 0 */
  amountSol: number;
}

export type OrderProvider = typeof OrderProvider[keyof typeof OrderProvider];


export const OrderProvider = {
  pump: 'pump',
  jupiter: 'jupiter',
} as const;

export interface Order {
  provider: OrderProvider;
  /** @nullable */
  requestId?: string | null;
  transaction: string;
  /** @nullable */
  outAmount?: string | number | null;
  /** @nullable */
  router?: string | null;
  /** @nullable */
  mode?: string | null;
  /** @nullable */
  feeBps?: number | null;
  /** @nullable */
  feeMint?: string | null;
  /** @nullable */
  expireAt?: string | null;
  /** @nullable */
  lastValidBlockHeight?: number | null;
}

export type ExecutionInputProvider = typeof ExecutionInputProvider[keyof typeof ExecutionInputProvider];


export const ExecutionInputProvider = {
  pump: 'pump',
  jupiter: 'jupiter',
} as const;

export interface ExecutionInput {
  signedTransaction: string;
  /** @nullable */
  requestId?: string | null;
  provider: ExecutionInputProvider;
}

export interface ExecutionResult {
  status: string;
  /** @nullable */
  signature?: string | null;
  /** @nullable */
  code?: number | null;
  /** @nullable */
  error?: string | null;
}

export type GetBalanceParams = {
address: string;
};

export type ValidateMintParams = {
mint: string;
};