# lib/api-zod/src/generated/types/balance.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Solana multi-wallet panel API
 * OpenAPI spec version: 1.0.0
 */

export interface Balance {
  address: string;
  lamports: number;
  sol: number;
}