# lib/api-zod/src/generated/types/errorResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Solana multi-wallet panel API
 * OpenAPI spec version: 1.0.0
 */

export interface ErrorResponse {
  error: string;
  /** @nullable */
  code?: number | null;
  /** @nullable */
  errorCode?: string | number | null;
  /** @nullable */
  router?: string | null;
  /** @nullable */
  signature?: string | null;
  /** @nullable */
  status?: string | null;
}