# lib/api-zod/src/generated/types/executionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Solana multi-wallet panel API
 * OpenAPI spec version: 1.0.0
 */
import type { ExecutionInputProvider } from './executionInputProvider';

export interface ExecutionInput {
  signedTransaction: string;
  /** @nullable */
  requestId?: string | null;
  provider: ExecutionInputProvider;
}