# lib/api-zod/src/generated/types/executionResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Solana multi-wallet panel API
 * OpenAPI spec version: 1.0.0
 */

export interface ExecutionResult {
  status: string;
  /** @nullable */
  signature?: string | null;
  /** @nullable */
  code?: number | null;
  /** @nullable */
  error?: string | null;
}