# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Solana multi-wallet panel API
 * OpenAPI spec version: 1.0.0
 */

export interface HealthStatus {
  status: string;
  rpc: string;
  jupiterConfigured: boolean;
}