# lib/api-zod/src/generated/types/mintInfo.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Solana multi-wallet panel API
 * OpenAPI spec version: 1.0.0
 */

export interface MintInfo {
  mint: string;
  valid: boolean;
  decimals: number;
  supply: string;
  program: string;
}