# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Solana multi-wallet panel API
 * OpenAPI spec version: 1.0.0
 */
import type { OrderProvider } from './orderProvider';

export interface Order {
  provider: OrderProvider;
  /** @nullable */
  requestId?: string | null;
  transaction: string;
  /** @nullable */
  outAmount?: string | number | null;
  /** @nullable */
  router?: string | null;
  /** @nullable */
  mode?: string | null;
  /** @nullable */
  feeBps?: number | null;
  /** @nullable */
  feeMint?: string | null;
  /** @nullable */
  expireAt?: string | null;
  /** @nullable */
  lastValidBlockHeight?: number | null;
}