# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Solana multi-wallet panel API
 * OpenAPI spec version: 1.0.0
 */

export interface OrderInput {
  wallet: string;
  outputMint: string;
  /** @exclusiveMinimum 0 */
  amountSol: number;
}